requirements.txt
discord.py==1.6.0
geopy==2.1.0
praw==7.2.0
feedparser==6.0.2
discord==1.0.1
python-dotenv==0.16.0
